Implement consistency check for FMI 1.0 continuous states
- Added `numberOfContinuousStates` to `ModelMetadata`.
- Updated FMI 1.0 and 2.0 checkers to extract state and event indicator counts.
- Implemented `checkContinuousStatesAndEventIndicators` in `Fmi1ModelDescriptionChecker` to verify that Model Exchange FMUs provide enough continuous variables for states, derivatives, and event indicators as required by the spec.
- Added negative test cases and test data for FMI 1.0.

+++ lib/checker/model_description/fmi2_model_description_checker.cpp
@@ -1594,7 +1594,7 @@ ModelMetadata Fmi2ModelDescriptionChecker::extractMetadata(xmlNodePtr root)
     {
         try
         {
-            metadata.numberOfEventIndicators = std::stoul(*num_event_ind);
+            metadata.numberOfEventIndicators = static_cast<uint32_t>(std::stoul(*num_event_ind));
         }
         catch (...)
         {
